Enable openvds version extraction from Dockerfile
As we plan to cache openvds image in github actions, we have to make
sure we cache correct version.

One option would have been to just make version a build argument (with
no default value). That would however lead to version string repeating
both in github actions and in radixconfig, which is prone to
inconsistency as one needs to remember to update in both places.

Chosen option is less nice, but will allow us to automatically extract
openvds version for github actions cache.