Initialize response struct values
While not exactly an error, this was convenient change when writing
memory tests. Valgrind complained that we are working on uninitialized
memory.

The change is useful in case when our c++ endpoint handling code throws
an error. Then response data is not set, but we still can call
response_delete safely.

Currently we call response_delete in go unconditionally.
We do not write tests through go, so it is not perfectly clear if that
change has any positive impact on server behavior, but it is possible.