Add memory tests suite
The best option would have been to do memory testing through go.
However:

- go supports asan (and thus lsan), doesn't support valgrind and has no
plans for it [1]

- not closing openvds resources is not considered a "proper" memory
leak (which means one has no chance to free the memory as there are no
pointers to that memory anymore).
Valgrind reports it in the "still reachable" category, which
usually is considered "everything is probably fine, you still have some
resources you didn't close, but you still can".
It looks like asan detects only proper memory leaks. At least I was
unable to detect forgotten-close issue through asan

- go doesn't support leak checks when c is called from go. To
achieve that one must explicitly call C.__lsan_do_leak_check() in the
code (so this might be even called while some variables are still not
destroyed)

- it is not allowed to call C in go test files, so we would need to
create proper go run files instead of using testing mechanism and
testing libraries.

- it seems like lsan errors can be intermittent. In some test code
memory leaks were/were not reported on sequential runs.

So trying to implement memory testing suite in go doesn't seem to be
worth it.

Due to that decision was made to test only c library with valgrind.
It is not optimal as we still don't control how we use c calls, but it
is this or nothing.

[1] https://github.com/golang/go/issues, issue 60600